Replace HashMap with LinkedHashMap to guarantee the order
We are using HashMap in UiccCard to map portIndex <-> UiccPort.
During download process, we expect card.getUiccPortList()[0] inside getFirstActiveEuiccPort API to always return the same UiccPort for each call.

mUiccPorts.values() returns collection of unordered so the order is not guaranteed which causes card.getUiccPortList()[0] problem.

So we have replaced HashMap with LinkedHashMap to guarantee the order.

Test: Manual, added debug logs in getFirstActiveEuiccPort and verified
it always returns same port in E+E mode
Bug: 237560274

Change-Id: I722a222d2b122da873e8bb9e9d604e09a6793c7f