canManageSubscriptionOnTargetSim should check all eligible active eSIM profiles
the current logic in canManageSubscriptionOnTargetSim does not handle the case of multiple active eSIM profiles and it will only check the first matched active profile. For the API calls like getEid, which does not use port index, the call might fail due to canManageSubscriptionOnTargetSim only check the first profile.

Test: 1. Manually Verified on P10, not observed SIM card Could not found
screen.
2. Device originally on P+E mode and users download activate Fi profile first -> can get Eid(PASS) -> users download another non-Fi profile -> MEP mode with multiple active embedded subscriptions-> Fi calling getEid again(PASS)
Bug: 236172214

Change-Id: I9dd31c7977fad3a660523b196a8cb8a08846264e