Always re-create the network agent in non-suspended state
Connectivity service does not support creating a network
agent in suspended state, and the network agent can't become
connected when suspended. So we need to make sure the agent
is created in non-suspended state, and then immediately
update to the correct suspended state.

Test: Manual + atest DataNetworkTest
Bug: 237335416
Merged-In: Ic0279f7b9dea831e1e00758f0048471dede7682e
Change-Id: Ic0279f7b9dea831e1e00758f0048471dede7682e