Move dataEnabled override rule check to phone switcher
Previously, whether an override rule applies to the current condition is
decided by each sub.

The motive for moving the check to phoneSwitcher is that as the future
override rule becomes increasingly dependent on the status of both SIM,
phoneSwitcher, which oversees both SIMs in order to route requests, is
more suitable.

Bug: 244064524
Test: atest + manual toggle dataDuringCall during voice call + toggle
always allow MMS for sending MMS.

Merged-In: I8ea07a1ed879afd52d960241b40e8a4897f899aa
Change-Id: I8ea07a1ed879afd52d960241b40e8a4897f899aa