Disable signal strength reporting when threshold is empty
This change optimizes scenarios that when the signal threshold
is empty, the SignalThresholdInfo is disabled before setting
signal strength reporting criteria to modem.

The issue can reproduce with fulfill of both of the conditions below:
- Device is in idle mode
- A customized SignalStrengthUpdateRequest was set with
shouldReportWhileIdle to true and shouldReportSystemWhileIdle
to false.

In above scenario, the system signal thresholds which were not
set by client may leave with empty threshold and enabled.
The optimization here makes sure only the client's non-empty
threshold is enabled.

Bug: 245658506
Test: atest  SignalStrengthControllerTest
Test: feature regress test (go/signalstrength-s-ft)
Change-Id: I22dceb4a3f34274942ca9a278d657c6d2263160c