Default to not mutating isOpportunistic bit
When the carrier config loads from a carrier app, if for some
reason the app doesn't have information about a specific
subscription it might not attempt to override the isOpportunistic
bit, but the current code actually resets the bit's value even if
the carrier doesn't attempt to change it. Rather than mutating
the value simply as the result of the carrier config loading, the
value should persist either as the default OR as the last value
that the carrier set it to. While this doesn't actually resolve
the root cause of the bug, it prevents a class of issues where the
carrier app doesn't have enough information from resulting in an
unwanted mutation of the existing state. In absence of information,
the safest thing to do is nothing.

Bug: 254291382
Test: atest SubscriptionInfoUpdaterTest#\
        testOpportunisticSubscriptionNotUnsetWithEmptyConfigKey
Change-Id: Ib3bfd37b575290fce506981fdc91970d49d2bc56