Transition to current state after icon timers expire
When the current icon changes temporarily to another state with a timer,
there is a chance that the icon will change again before the transition
completes. This causes NetworkTypeController to preemptively cancel the
transition timer but not update the current state. The race condition
will be fixed in b/233048459, but add a transitionToCurrentState() after
clearing timers to make sure the state is accurate.

Test: atest NetworkTypeControllerTest
Test: manually test basic timer functionality
Bug: 252734651
Change-Id: I260b95ac1c074777d293d387d608596cda4599b1