Correct the value of cookie
The cookie in EVENT_AIDL_PROXY_DEAD is used to check if the death
notification should be ignored. The cookie object is passed by reference which causes the msg.obj.get() always equal to mServiceCookies.get(aidlService).get() when handling EVENT_AIDL_PROXY_DEAD. So pass the long type value of cookie in EVENT_AIDL_PROXY_DEAD message. And since the service is RADIO_SERVICE when handling EVENT_RADIO_PROXY_DEAD, the cookie of RADIO_SERVICE should be used for linkToDeath() when radio version is less than 2.0.

Bug: 258390132
Test: Manual test by force killing RILD
Change-Id: I7087f01fb56a611c21a2f976b05b85f67dafa50b