Refactor SMS handling for domain selection
This change is to apply the domain selection model.
If the domain selection model is supported, the domain selection for
sending a MO SMS will be done by the domain selection service and
SmsDispatchersController should wait for the completion of domain
selection before sending a MO SMS. SmsDispatchersController sends a MO
SMS according to the selected domain - DOMAIN_PS uses ImsSmsDispatcher,
and DOMAIN_CS uses CdmaSMSDispatcher or GsmSMSDispatcher.
If the domain selection model is not supported, it works with the
existing logic to provide backward compatibility.

The followings were considered while refactoring this component:
1) interwork with the domain selection module asynchronously.
2) interwork with the EmergencyStateTracker if an emergency SMS.
3) should handle sending the multiple SMS while the domain selection is
   running.
4) should handle sending the retry SMS also.

Bug: 243377249
Test: atest SmsDispatchersControllerTest
Test: manual (verify IMS/GSM/CDMA MO SMS for normal/emergency SMS when domain selection disabled)
Test: manual (verify retry SMS for normal/emergency SMS when domain selection disabled)
Test: manual (verify IMS/GSM/CDMA MO SMS for normal/emergency SMS when domain selection enabled)
Test: manual (verify retry SMS for normal/emergency SMS when domain selection enabled)

Change-Id: I09d9af248d6bee6a2ae50b408d07cad4362a57ba