Honor Feature Flag When Updating Null Cipher Settings
When a GsmCdmaPhone instance is asked to handle a null cipher
update, it will first check the cellular_security namespace for
the enable_null_cipher_toggle flag. If the flag is false (the default),
no command will be sent through the RIL to the modem.

When the feature flag is set to false after previously having been set
to true (feature disablement), the user's device will maintain the
behavior of the active setting until the modem reboots. Since one reason we may want to disable this feature is an issue in the HAL integration itself, it could be unsafe to try and call the modem to reset state, so we implement the simpler and safer option.

Bug: b/262063470
Test: atest GsmCdmaPhoneTest
Change-Id: I0ab9f118ef2df816c3b878c3a34d60f1c35518ab