Fix a test failure reported as flaky
When the BroadcastReceiver is used for receiving the sent result, but it
can be reported as a flaky test since sometimes broadcasts can take a
long time if the system is under load.
This change uses the IActivityManager instead of the BroadcastReceiver
to check whether the PendingIntent#send(int) is successfully called or
not.

Bug: 262934689
Test: atest FrameworksTelephonyTests, SmsDispatchersControllerTest
Change-Id: I1111e7c0f329080149572fb7484f25ea8b9397ac