Update override network type after service state changes
Previously, if the data rat was the same after a service state change,
we would not do anything since the state is the same.
However, the override network type could be different.
Make sure to update the override network type (either directly or
through a transition) whenever the service state changes.

Test: atest NetworkTypeControllerTest
Bug: 263678417
Change-Id: I74fb8f483a2dec281d16f86bf0f16e19acae9221