Track Modem Support for Null Cipher Disablement in Phone
Adds Phone.isNullCipherAndIntegritySupported which,
by default, is initialized to false in Phone (no support).

When the radio becomes available, it attempts to make
a request to set the user's null cipher and integrity
preferences. If the response is anything other than a
REQUEST_NOT_SUPPORTED exception, the phone then considers
the feature supported by the modem.

Bug: 264574296
Test: atest GsmCdmaPhoneTest
Change-Id: Ib3ba1aa28da35758dbdf585704d929c9bc45ff97