Ensure We Never Serialize Bad allowed_network_types_for_reason Keys Into Sim DB
Change the default case to throwing an exception. Better to not
silently fail and corrupt the DB.

Add test cases for loading network types from the SIM DB, including
serialization and deserialization. The cases cover both the existence of
bad keys and values in the DB and ensure we ignore those on loadr. and the
assurance that we will not write an unknown value into the SIM DB.

Bug: 266472206
Test: atest GsmCdmaPhoneTest
Test: Manual testing; force a bad value into the sim db and ensure we
don't read it back.(adb root; adb shell content query --uri content://telephony/siminfo)

Change-Id: I62ae213d75047b8abd0469ee6e12c5d4fa43b539