Allow mutable implicit PendingIntent
Starting from target SDK U, creation of mutable PendingIntents with implicit Intents will be blocked for security reasons. This test creates such an object and fails, so a quick mitigation is to apply a flag PendingIntent.FLAG_ALLOW_UNSAFE_IMPLICIT_INTENT that bypasses the above block.

The flag is introduced in ag/21018466.

Bug: 266799056
Test: atest com.android.internal.telephony.SmsDispatchersControllerTest
Change-Id: I94595c7dbc6c61cc36c33b7dda963d7ac7a93f1d