Fix testInitializationFromDb
This test was previously passing without the correct permissions, since it just so happened that the instrumentation target's PhoneFactory wasn't initialized.

Instead of relying on the instrumentation context directly, rely on the
TelephonyTest ContextFixture, which already handles privileged phone state read permissions, which are expected by the internal telephony framework.

Bug: 267186351
Fixes: 267186351
Test: atest GsmCdmaPhoneTest InboundSmsTrackerTest
Change-Id: I323d11f84726a403af7af71399ad01d62c52e517