Fix a potential issue with the state machine of EuiccConnector
When the service is dies, the binder is disconnected and the state is transition to the Disconnected state.
If the service is not restarted for a period of time afterward, it will transition to the Available state.
Fixed an issue that prevented service_connected from being handled if the service is restarted in the Available state

Bug: b/270110619
Test: atest EuiccConnectorTest
Change-Id: Idef5dfb024f6c7ad82b29060d58d549d99531182