Fix phone process crash due to NPE in DataNetwork
This CL calls StateMachine#sendMessage instead of Message#sendToTarget
to avoid the NPE when the handler of the message is null.

The StateMachine handler could be null when e.g. it has quit. In
this case, calling Message#sendToTarget will cause NPE and crash
the phone process. StateMachine#sendMessage has protected against
the case to avoid the NPE when handler is null.

Bug: 270902239
Test: atest DataNetworkTest
Change-Id: Ida123e2b493cb70b005b2ac47843d7c6d6f14656