Ignore ECBM enter/exit event on Phone when domain selection supports
Basically, the ImsService/modem should not be sending the ECBM enter
event when domain selection is enabled. However, if the ImsService/modem
sends the ECBM enter event, the framework may malfunction.
In order to avoid this unexpected behavior, the existing implementation
for ECBM enter/exit event handling will be ignored on
ImsPhone/GsmCdmaPhone when domain selection is enabled. Instead, this
ECBM enter/exit event will be handled by the EmergencyStateTracker when
domain selection is enabled.

Bug: 270674175
Test: atest GsmCdmaPhoneTest, ImsPhoneTest
Change-Id: I589bc27046e1b1633548e0f13c5c9efed88922aa