Force update override network type before timer
When data RAT changes from LTE -> NR but the state is no longer NR
advanced, we don't update the override network type if a timer is
active. For this scenario, manually update the override network type to
be NONE to prevent a mismatch between the data rat and override network
type.

Test: atest NetworkTypeControllerTest
Bug: 269163800
Change-Id: I686fa530814d522ed0b1216c63baf21e1fd00cba