Use "useEmergencyPdn" flag for emergency SMS
This flag was introduced to check whether IMS PDN should be used for PS
emergency services. If IMS PDN is used for an emergency SMS, the domain
selection should not try to change the preferred transport type for
emergency service.
If true, the preferred transport type for emergency service will be
requested to the QualifiedNetworksService when the transport type needs
to be changed. Otherwise, the preferred transport type will not be
requested to that component even if the transport type needs to be
changed.

Bug: 270333880
Test: atest SmsDomainSelectionConnectionTest, EmergencySmsDomainSelectionConnectionTest
Test: manual (verify normal/emergency SMS on device)
Change-Id: I252f8e7aeb990cc62757a9f5efd40cb341e1ef49