Allow IWLAN network when power off modem
Since IWLAN network shouldn't be affected by modem, this change keeps
the networks that are established on IWLAN when modem power off.

The IMS registration change is to track which specific network where the IMS/RCS registration took place, so that fwk only wait for dereg delay for those networks. Without this change, if APM is on, fwk always wait for an extra delay for a network that's requested by IMS/RCS service but irrelevant to registration.

Fix: 256025291
Test: Confirm regular voice call drop upon APM ON but WFC is preserved
Test: b/279801130#comment16, P22+P23 IMS related valid test pass
Change-Id: I61be3c79566a386675fe1e9c36d87dd70018c1ba