Ensure MERGE_COMPLETE connection event propagates for IMS conferences.
The connection event MERGE_COMPLETE was not propgating nor getting logged to the dumpsys as it was being invoked after the original calls had been diconnected in favor of the conference call and thus the listeners were removed at that point. This change ensures the MERGE_COMPLETE connection event properly propagates and logs to the dumpsys before both the original background call and foreground call are disconnected in favor of the conference call for IMS conferences.

Bug: 265187856
Test: Manually create IMS conference call. Ensure MERGE_COMPLETE logs to the dumpsys for each of the original calls at time of merge.
Change-Id: I2f3b9759c987434afa837863a9cebab16521fb55