Remove AnomalyReporter code
AnomalyReporter is called if UiccSlotInfo is null when getting slot id using card id.
UiccSlotInfo will be set in the array by the number of slots and set the value of UiccSlotInfo in the array when the slot_status_done occur.
If that value is read before slot_status_done, the array will not be null but its value may be.
In this case, return the slot id as an invalid_slot_index, so we remove the AnomalyReporter we were using.

Bug: b/260469451
Test: atest EuiccConnectorTest
Change-Id: I726d374e679a77898a86a189c1e223459accc511