Fix quick response not sending via work profile.
Telecom calls into SmsController to send the SMS message over
the work sim. However, the calling user is always User 0,
which results in the message failing to send since the work sim is not
associated with the primary profile.

To reduce the scope of impact to other APIs, we will bypass the
validation check only in SmsController#sendTextForSubscriber if the
calling user has granted the INTERACT_ACROSS_USERS_FULL permission.
Adding log (debug level) to track when validation is bypassed.

Fixes: 275115413
Test: Manual (can see that the text message was sent to caller)
Change-Id: Ie992df943b452c85aebc816b0c85145426c24201