Notify link status changed on connected or disconnected only
NetworkTypeController relies on link status changed events to
distinguish RRC idle vs connected if it does not use physical channel
config indications.
If data is unexpectedly disconnected, DN/DNC might not send the link
status changed event. When a DataNetwork disconnects, update the link
status to make sure we don't miss an INACTIVE state update.
When we receive a data call response in connecting state, the link
status can be prematurely updated. Update the link status only and
defer notifying DNC of the change until the data network is in
connected or disconnected state.
If the data network is already in connected state, update the link
status if it has changed.

Test: atest DataNetworkTest, DataNetworkControllerTest
Test: manual basic data tests
Bug: 279380751
Bug: 276710123
Change-Id: I877b3ed736a0c81b8a0dd692dfdc49a832642ab0