Ensure ScanResult is for Active Request
Even after a scan result is no longer active, due to
the fact that request and response paths are separate,
it is possible to still receive scan results occasionally
for requests that are no longer active. While this
technically represents a bug at the lower layers, it would
be impractical to entirely close that race condition. Hence
the NetworkScanRequestTracker, which is the ultimate source
of truth for this synchronization, should do the check.

Introduce the UT to reproduce issue and verify it is fixed
with the patch.

Bug: 282921687
Test: atest NetworkScanRequestTrackerTest
Change-Id: Ia4953a4bd6bdded35231ba88b46df5626802dc62