Remap CODE_SIP_BAD_REQUEST to SERVER_ERROR for outgoing calls.
The vendor IMS stack can report a number of disconnect causes for
incoming calls which were rejected on the device by the lower levels and
as a consequence become "auto missed calls".
CODE_SIP_BAD_REQUEST is one of the codes that can be reported which we
will attribute as an "auto missed call".
However, it is ALSO possible for an outgoing call to return that same
disconnect code, which we should not remap to the auto rejected cause.
Added a check to ensure that if the call state is dialing or alerting we
will remap to SERVER_ERROR rather than the auto-missed disconnect code.
This ensures failed outgoing calls don't get treated as missed incomign
calls.

Fixes: 281320323
Test: Added unit test for disconnect cause remapping.
Change-Id: Ib272858cca7f6bd1694183e7f090fe1cdbe45b27