Add signal strength callback
The existing telephonyCallback for signal strength is registered by sub Id, so the registrant cares about the phone id only, it will need to monitor the sub Id change on this phone slot. Adding a sub id agnostic callback solves this issue, which is the behavior that's consistent with DisplayInfoController.

Bug: 260928808
Test: phone call + internet browsing
Test: daily test P22 https://testtracker.googleplex.com/efforts/testplans/1498577 and P23 https://testtracker.googleplex.com/efforts/testplans/1498694
Change-Id: I1c8f1b6fec2674be62571e7354510085fda10e59
Merged-In: I1c8f1b6fec2674be62571e7354510085fda10e59