Check default APN type before set preferred profile
Without the change, a DUN | DEFAULT data profile will be updated to be the default if it's the only connected internet network at the time even though a preferred data profile exists. This happens when single connection rule is in place. To prevent setting DUN | DEFAULT profiles as the preferred when there exists a clean DEFAULT profile, this change assumes an existing preferred profile, if any, has the right APN types, and thus future preferred profile candidates should resemble.

Fix: 287171389
Test: basic testing +  reviewed log + verified by reporter
Change-Id: I3d69485ee9343d5968dff5a3ec64f4d74bc88d98