Add logic to keep 5G+/5G UC icon when RRC idle
Some carriers use an empty PCC list to indicate RRC idle, but the issue
is that we also use the PCC list to determine whether the device is
NR advanced to show the 5G+/5G UC icon.
Add logic to ratchet the bands and bandwidths from the PCC if the anchor
NR cell is the same so we can keep the NR advanced state, and only
update the values if the anchor NR cell changes.
Then we can handle 5G+ when RRC idle and prevent 5G/5G+ flickering.

Test: atest NetworkTypeControllerTest
Test: manual verify 5G+ shown when RRC idle
Bug: 288607222
Change-Id: Ie7f76a87851021cfb84925f365451c240c0b0280