Check user settings only for data during call
During the phone call, we should stick with the voice phone regardless whether data can be established on it. This is consistent with pre-U behavior.

Without this change, if EMC is happening on nDDS, but data is currently OOS, then we don't switch to the nDDS. However, if later on the EMC nDDS is in roaming, then we don't have another chance to resurrect data service on this nDDS.

Fix: 293221772
Test: use data during voice call using command to reproduce scenario
Change-Id: I15f05ec2f4878c4f0b501caa36582fb35b55e2d3