telephony: RadioMessagingProxy: Fix sending CDMA SMS in sendImsSms()
When cdmaPdu is not null, we need to set the message technology to THREE_GPP2
so that dispatchImsCdmaSms() is chosen in libril. If we don't do this, libril
chooses dispatchImsGsmSms() instead, which means that the message is empty if
AIDL is being used and may be empty or invalid if HIDL is being used.

Change-Id: If287c6fa1712a624531e11dd09aeaf7a6c7914c6