Delay data switch untill call alerting
For emergency call placed on non-DDS when APM is on, modem expects an order of
commands: first the emergency call, then dds switch. If the order were reversed due to race condition, modem will abort the emergency call.

The change exclude the Dialing call state as the switch trigger, to give
modem time to deal with the EMC first.

Fix: 298291981
Fix: 284412852
Fix: 299585246
Test: Verified at b/298291981#comment82
Test: Daily test verified in b/299355154#comment5
Change-Id: Ia200072a44d6141ab4830d2143ae0483fa653856