Adds request and response for handling network validation feature.
A new request network validation function has been added. Only QNS
can request this, and a network validation can be requested to each
DataNetwork through AccessNetworksManager. DataNetwork can request
validation to the DataServices such as Cellular or IWLAN through the
DataService API. If there is a DataService that does not support the
network validation feature, it may reply UNSUPPORTED as a default
result.
The Validation status variable has been added to DataCallResponse in
response to a network validation request. Network validation status
has also been added to PrecisedataConnectionState so that privileged
apps can use this variable to know.

Bug: 299346675
Test: atest FrameworksTelephonyTests
Change-Id: I77217967731ccadf9efdfe918d773ae8994f49d0