Add interworking with EmergencyStateTracker for sending emergency SMS
This adds the interworking implementation for emergency SMS with the
EmergencyStateTracker. When an emergency SMS is being sent, the
SmsDispatchersController calls the startEmergencySms, then the
EmergencyStateTracker sets the emergency mode to the modem and if the
emergency SMS completely sends regardless of the success or failure, it
should call the endSms, then the EmergencyStateTracker exits the
emergency mode or enters the emergency SMS callback mode if needed.

Bug: 262804071
Test: atest SmsDispatchersControllerTest, EmergencyStateTrackerTest
Test: manual (verify emergency SMS using test number on device when
domain selection disabled)
Test: manual (verify emergency SMS using test number/equipment on device
when domain selection enabled)
Test: manual (verify normal MO SMS on device when domain selection
disabled)
Test: manual (verify normal MO SMS on device when domain selection
enabled)

Change-Id: I7310c1affa3d8e474de7f3f9f81a9af56a17c2c9