Handle CellularIdentifierDisclosure events in telephony
This CL changes the previous noop logic in the RIL for handling
`cellularIdentifierDisclosed` calls from the modem.

The RIL now exposes APIs to register and deregister for disclosure
events. To avoid AIDL-based instances of CellularIdentifierDisclosure
propagating through the framework, disclosure objects are converted
to Java wrappers before notifying registrants.

We also define the first registrant for disclosures,
`CellularIdentifierDisclosureNotifier`. For now, this class is mostly a
stub and does nothing but log events. It's implementation will be
completed in future CLs. Each `Phone` instance holds on to
a reference to the singleton notifier object, and passes the disclosure
event on to the notifier when it is received.

Both initializing a `CellularIdentifierDisclosureNotifier` and
registering for `cellularIdentifierDisclosed` events are guarded by a
new feature flag.

Test: atest GsmCdmaPhoneTest
Bug: 309955068
Change-Id: I10b1f2c07c1fad82e3b7e7bbd08198aeed49ea49