Update modems on identifier disclosure transparency preference changes
The telephony service delegates to each Phone instance to inform the
modem of user preferences. I.e. when the user wants this
transparency feature, we enable the functionality in the modem. When the
user no longer wants the feature, we disable the functionality.

When the radio becomes available, we make sure to tell the modem whether
or not we'd like identifier disclosure transparency enabled. This is
needed because modems are not expected to persist state across reboots.
We also use the response to cache whether or not the modem supports this feature.
That way, we can respond to callers with an exception without calling
into the modem (faster failures).

If the modem supports this feature, we update the modem when a
caller lets us know state has changed.

Bug: 313475601
Test: atest GsmCdmaPhoneTest
Change-Id: Ib4719f12c61b09466a770496855d4d382eebc4c9