Create WAP cache and API for retrieving WAP message sizes.
As part of the effort to support MMS over satellite, we want to limit the size of MMS that can be downloaded while using a satellite connection. This cl implements a cache of message sizes retrieved from WAP push messages. In a future cl, the message size will be retrieved from the cache during mms download and if it exceeds some threshold the download will be prevented.

See the design of the cache here go/android-wap-push-cache.

Test: atest WapPushCacheTest SmsControllerTest
Bug: 311244479

Change-Id: Ie5fc3b1043141a8d42a6f6f08d99e86525f20b20