Fix minimum ril hal version required by getSystemSelectionChannels
In radio proxy, getSystemSelectionChannels only works in version 1.6
or higher. Currently, versions 1.4 and 1.5 should respond that RIL
does not support the request, but RIL was unable to do so.
As a result, in versions 1.4 and 1.5, requests are made but remain
unanswered. Change this to give a response that the request cannot
be supported in versions 1.4 and 1.5.

Bug: 314709813
Test: CtsTelephonyTestCases
Change-Id: I891fbf511d9767101aa295c2333906d7e9013516