Handle security algorithm updates
Turn the HAL callback into a `Phone` event for those who register for
it. Register `GsmCdmaPhone` for the event and handle it by forwarding to
the `NullCipherNotifier`.

The notifier justs logs the event at the moment, but it will later
aggregate the events and publish derived events about the use of null
ciphers.

Bug: 315005938
Test: atest SecurityAlgorithmUpdateTest GsmCdmaPhoneTest
Change-Id: I9b4e7e817dd28ea456071bb27bb675f3ac923f8e