Update modems on null cipher notification preference changes
This commit adds the APIs to the Phone class to support enabling /
disabling null cipher notification in the modem. When the radio
becomes available, the modem is notified of the user preference
for null cipher notification as the modem is not expected to
persist state across reboots. The response as to whether the modem
supports this feature is then cached to allow for faster failures
without needing to call into the modem.

Bug: 316592273
Test: atest GsmCdmaPhoneTest
Change-Id: I4354bb27302adc4695dc5a699f467b202ab33a29