Don't send setN1Mode for partial CC
Enabling and disabling N1Mode is a heavyweight
operation at the modem, which can result in OTA
signalling. In order to reduce the churn caused by
the SIM card loading and unloading, particularly
during a SIM reset procedure (even though this isn't
a common occurrence, but also to address a carrier
test), we can avoid making any transient changes
that occur when the CarrierConfig changes due to
no SIM being present.

Bug: 319522224
Test: atest GsmCdmaPhoneTest#\
          testNrCapabilityChanged_firstRequest_incom\
          pleteCarrierConfig_changeNeeded
Test: telephony sanity (voice, data, sms)
Change-Id: Ief07f27c7ccaa2234e2c309fcc11be36169fafa8