Call setN1Mode via Phone rather than CommandsInterface
Update the ImsNrSaModeHandler to control the N1Mode via the
Phone object (which will implicitly call the GsmCdmaPhone object)
instead of going directly to the commands interface. This will
ensure that the desired value of N1 is properly combined in
a single location and avoid some unlikely edge cases that would
cause it to be out of sync.

Bug: 319936478
Test: Telephony Sanity (voice, data, sms) on T-Mobile
Test: Manual verification of new+old code paths
Change-Id: Ie8c9f249e40c157d10c853caa28fe258aec6ef29