Protect against NPE in onCallQualitychange callback.
There are cases where a call session is already cleaned up for a call and
the vendor IMS stack reports a change in call quality after that.  This is
most likely due to timing issues.

Add an NPE check for ImsCall#getSession; worst case we are going to drop
a call quality report -- this is fine since the call session has been
cleaned up so the call quality is moot.

Flag: Simple NPE fix to go into next release.
Fixes: 321133325
Test: Compile; run unit tests.
Change-Id: I230ce5d7e8a8c9787801a5e9919d1288df0bf1d2