Add new api for satellite entitlement status  and satellite entitlement plmns
Added an API to find whether the satellite entitlement status is enabled by the entitlement query result.
Added an API to check the satellite entitlement plmns based on the entitlement query result.

Bug: 322143408
Test: atest SubscriptionInfoInternalTest, SubscriptionDatabaseManagerTest
Test: manual test
1. onCreate case
verified the COLUMN_SATELLITE_ENTITLEMENT_STATUS value set to 0.
verified the COLUMN_SATELLITE_ENTITLEMENT_PLMNS value set to empty.
2. onUpgrade case
verified the COLUMN_SATELLITE_ENTITLEMENT_STATUS value did not exist, but was set to 0 after the upgrade.
verified the COLUMN_SATELLITE_ENTITLEMENT_PLMNS value did not exist, but was set to empty after the upgrade.

Change-Id: I1e7cfcd283ff442a7867f933d63b046894a9e4b1