Store the Satellite Entitlement Status and  the Satellite Entitlement plmn list in db and added restrict reason for entitlement per carrier according to this value
If this carrier supports entitlement and the cache value or the Entitlement status value read from the db is disabled, add a restrict reason before the entitlement query.
If this carrier supports entitlement and the cache value or the Entitlement plmn list value read from the db is not empty, used it before the entitlement query.

Store the entitlement status and entitlement plmn list in the db after the query is completed.

Bug : 321876569
Test: atest SatelliteControllerTest
Test : manual test
1. If the entitlement status read from cache or db is disabled, verify adding a restriction reason before the entitlement query.
2. If the entitlement plmn list read from cache or db is not empty, verify for use it as plmn list before the entitlement query.
3. After the query is completed, verify storing the entitlement status and entitlement plmn list in the db.

Change-Id: If873463f560da3c188fa9e00b002d77edde0303d